+++ src/main/kotlin/de/achimonline/ansible_lint/annotator/AnsibleLintAnnotator.kt
@@ -15,24 +15,30 @@ import com.intellij.openapi.options.ShowSettingsUtil
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiFile
+import de.achimonline.ansible_lint.annotator.AnsibleLintAnnotator.*
 import de.achimonline.ansible_lint.annotator.actions.AnsibleLintAnnotatorClipboardAction
+import de.achimonline.ansible_lint.annotator.actions.AnsibleLintAnnotatorIgnoreFileAction
 import de.achimonline.ansible_lint.annotator.actions.AnsibleLintAnnotatorNoQAAction
 import de.achimonline.ansible_lint.annotator.actions.AnsibleLintAnnotatorOpenUrlAction
 import de.achimonline.ansible_lint.command.AnsibleLintCommandLine
 import de.achimonline.ansible_lint.parser.AnsibleLintItem
 import de.achimonline.ansible_lint.parser.AnsibleLintParser
 import de.achimonline.ansible_lint.bundle.AnsibleLintBundle.message
-import de.achimonline.ansible_lint.command.AnsibleLintConfigFile
+import de.achimonline.ansible_lint.command.file.AnsibleLintCommandFileConfig
+import de.achimonline.ansible_lint.command.file.AnsibleLintCommandFileIgnore
 import de.achimonline.ansible_lint.common.AnsibleLintHelper
 import de.achimonline.ansible_lint.notification.AnsibleLintNotification
 import de.achimonline.ansible_lint.settings.AnsibleLintConfigurable
+import de.achimonline.ansible_lint.settings.AnsibleLintSettings
 import de.achimonline.ansible_lint.settings.AnsibleLintSettingsState
+import java.io.File
 import java.io.IOException
 import java.util.concurrent.ExecutionException
+import kotlin.io.path.pathString
 
 private val LOG = logger<AnsibleLintAnnotator>()
 
-class AnsibleLintAnnotator : ExternalAnnotator<AnsibleLintAnnotator.CollectedInformation, List<AnsibleLintItem>>() {
+class AnsibleLintAnnotator : ExternalAnnotator<CollectedInformation, ApplicableInformation>() {
     private val ansibleLintNotification: AnsibleLintNotification = AnsibleLintNotification()
 
     override fun collectInformation(file: PsiFile, editor: Editor, hasErrors: Boolean): CollectedInformation {
@@ -43,15 +49,18 @@ class AnsibleLintAnnotator : ExternalAnnotator<AnsibleLintAnnotator.CollectedInf
         )
     }
 
-    override fun doAnnotate(collectedInformation: CollectedInformation): List<AnsibleLintItem> {
-        if (collectedInformation.hasErrors) return emptyList()
+    override fun doAnnotate(collectedInformation: CollectedInformation): ApplicableInformation {
+        if (collectedInformation.hasErrors) return ApplicableInformation()
 
         val project = collectedInformation.editor.project!!
         val settingsState = ApplicationManager.getApplication().getService(AnsibleLintSettingsState::class.java)
+        val configFile = AnsibleLintCommandFileConfig(project).locate()
 
-        if (settingsState.settings.onlyRunWhenConfigFilePresent && !AnsibleLintConfigFile.exists(project)) {
-            return emptyList()
-        }
+        if (settingsState.settings.onlyRunWhenConfigFilePresent && configFile == null) return ApplicableInformation()
+
+        val projectBasePath = AnsibleLintHelper.getProjectBasePath(project)
+        val relativeFilePath = collectedInformation.file.virtualFile.toNioPath().pathString.removePrefix(projectBasePath).removePrefix(File.separator)
+        val ignores = AnsibleLintCommandFileIgnore(project).parse()
 
         /**
             NOTE:
@@ -63,22 +72,27 @@ class AnsibleLintAnnotator : ExternalAnnotator<AnsibleLintAnnotator.CollectedInf
             - [Only trigger externalAnnotator when the file system is in sync](https://intellij-support.jetbrains.com/hc/en-us/community/posts/115000337510-Only-trigger-externalAnnotator-when-the-file-system-is-in-sync)
          */
         val tempFolderAndFile = AnsibleLintHelper.createTempFolderAndFile(
-            project,
-            collectedInformation.file,
-            collectedInformation.editor.document.text
+            projectBasePath = projectBasePath,
+            file = collectedInformation.file,
+            content = collectedInformation.editor.document.text
         )
 
         try {
             val lintProcess = AnsibleLintCommandLine(settingsState.settings).createLintProcess(
-                project.basePath!!,
-                tempFolderAndFile.first.path,
-                tempFolderAndFile.second.path
+                workingDirectory = projectBasePath,
+                projectDirectory = tempFolderAndFile.first.path,
+                configFile = configFile?.absolutePath,
+                yamlFilePath = tempFolderAndFile.second.path
             )
 
             val output = AnsibleLintCommandLine.getOutput(lintProcess)
 
             if (AnsibleLintCommandLine.SUCCESS_RETURN_CODES.contains(lintProcess.exitValue())) {
-                return AnsibleLintParser.parse(output.first)
+                return ApplicableInformation(
+                    settings = settingsState.settings,
+                    lintItems = AnsibleLintParser.parse(output.first),
+                    lintIgnores = ignores[relativeFilePath]?.toSet() ?: emptySet()
+                )
             } else {
                 LOG.error("Unexpected exit code from 'ansible-lint': ${lintProcess.exitValue()}; ${output.first}|${output.second}")
 
@@ -108,35 +122,41 @@ class AnsibleLintAnnotator : ExternalAnnotator<AnsibleLintAnnotator.CollectedInf
             }
         }
 
-        return emptyList()
+        return ApplicableInformation()
     }
 
-    override fun apply(file: PsiFile, annotationResult: List<AnsibleLintItem>, holder: AnnotationHolder) {
-        annotationResult.forEach {
-            if (it.getLine() != Int.MIN_VALUE) {
-                val document: Document = PsiDocumentManager.getInstance(file.project).getDocument(file)!!
+    override fun apply(file: PsiFile, applicableInformation: ApplicableInformation, holder: AnnotationHolder) {
+        applicableInformation.lintItems.forEach { lintItem ->
+            val itemIsIgnored = applicableInformation.lintIgnores.contains(lintItem.check_name)
 
-                val line: Int = it.getLine() - 1
-                val startOffset: Int = document.getLineStartOffset(line)
-                val endOffset: Int = document.getLineEndOffset(line)
-                val text = document.getText(TextRange(startOffset, endOffset))
-                val startOffsetDelta = if (it.getColumn() != Int.MIN_VALUE) it.getColumn() else (text.length - text.trimStart().length)
+            // skip ignored items; if specified via settings
+            if (!applicableInformation.settings.visualizeIgnoredRules && itemIsIgnored) return@forEach
 
-                holder.newAnnotation(
-                    getAnnotationHighlightSeverity(it),
-                    """
-                        ${it.description}
-                        ${if (it.content.body.isNotEmpty()) "(${it.content.body})" else ""}
-                        |
-                        Rule: ${it.check_name}
-                    """.trimIndent()
-                )
-                    .range(TextRange(startOffset + startOffsetDelta, endOffset))
-                    .withFix(AnsibleLintAnnotatorOpenUrlAction(it.url))
-                    .withFix(AnsibleLintAnnotatorNoQAAction(line, it.check_name))
-                    .withFix(AnsibleLintAnnotatorClipboardAction(it.check_name))
-                    .create()
+            // skip items with no line information
+            if (lintItem.getLine() == Int.MIN_VALUE) return@forEach
+
+            val document: Document = PsiDocumentManager.getInstance(file.project).getDocument(file)!!
+
+            val line: Int = lintItem.getLine() - 1
+            val startOffset: Int = document.getLineStartOffset(line)
+            val endOffset: Int = document.getLineEndOffset(line)
+            val text = document.getText(TextRange(startOffset, endOffset))
+            val startOffsetDelta = if (lintItem.getColumn() != Int.MIN_VALUE) lintItem.getColumn() else (text.length - text.trimStart().length)
+
+            var annotationBuilder = holder.newAnnotation(
+                if (itemIsIgnored) HighlightSeverity.WEAK_WARNING else getAnnotationHighlightSeverity(lintItem),
+                getAnnotationMessage(lintItem, itemIsIgnored)
+            )
+                .range(TextRange(startOffset + startOffsetDelta, endOffset))
+                .withFix(AnsibleLintAnnotatorOpenUrlAction(lintItem.url))
+                .withFix(AnsibleLintAnnotatorNoQAAction(line, lintItem.check_name))
+                .withFix(AnsibleLintAnnotatorClipboardAction(lintItem.check_name))
+
+            if (!itemIsIgnored) {
+                annotationBuilder = annotationBuilder.withFix(AnsibleLintAnnotatorIgnoreFileAction(lintItem.check_name))
             }
+
+            annotationBuilder.create()
         }
     }
 
@@ -155,9 +175,25 @@ class AnsibleLintAnnotator : ExternalAnnotator<AnsibleLintAnnotator.CollectedInf
         return HighlightSeverity.INFORMATION
     }
 
+    fun getAnnotationMessage(ansibleLintItem: AnsibleLintItem, ignored: Boolean): String {
+        return """
+            ${if (ignored) "${message("annotation.ignored-prefix")} " else ""}
+            ${ansibleLintItem.description}
+            ${if (ansibleLintItem.content.body.isNotEmpty()) "(${ansibleLintItem.content.body})" else ""}
+            |
+            Rule: ${ansibleLintItem.check_name}
+        """.trimIndent()
+    }
+
     data class CollectedInformation(
         val editor: Editor,
         val file: PsiFile,
         val hasErrors: Boolean
+    )
+
+    data class ApplicableInformation(
+        val settings: AnsibleLintSettings = AnsibleLintSettings(),
+        val lintItems: List<AnsibleLintItem> = emptyList(),
+        val lintIgnores: Set<String> = emptySet()
     )
 }