+++ src/main/kotlin/de/achimonline/ansible_lint/annotator/actions/AnsibleLintAnnotatorActionOrder.kt
@@ -1,7 +1,0 @@
-package de.achimonline.ansible_lint.annotator.actions
-
-enum class AnsibleLintAnnotatorActionOrder(val value: Int) {
-        OPEN_URL(1),
-        NO_QA(2),
-        CLIPBOARD(3)
-}