+++ src/main/kotlin/de/achimonline/ansible_lint/annotator/actions/AnsibleLintAnnotatorClipboardAction.kt
@@ -1,6 +1,6 @@
 package de.achimonline.ansible_lint.annotator.actions
 
-import com.intellij.codeInsight.intention.IntentionAction
+import com.intellij.codeInsight.intention.PriorityAction.Priority
 import com.intellij.ide.ClipboardSynchronizer
 import com.intellij.ide.CopyPasteManagerEx
 import com.intellij.openapi.editor.Editor
@@ -9,23 +9,7 @@ import com.intellij.psi.PsiFile
 import com.intellij.util.ui.TextTransferable
 import de.achimonline.ansible_lint.bundle.AnsibleLintBundle.message
 
-class AnsibleLintAnnotatorClipboardAction(private val ruleId: String) : IntentionAction {
-    override fun startInWriteAction(): Boolean {
-        return false
-    }
-
-    override fun getText(): String {
-        return "${AnsibleLintAnnotatorActionOrder.CLIPBOARD.value}: ${message("action.copy-rule-id-to-clipboard")}"
-    }
-
-    override fun getFamilyName(): String {
-        return "AnsibleLint"
-    }
-
-    override fun isAvailable(project: Project, editor: Editor?, file: PsiFile?): Boolean {
-        return true
-    }
-
+class AnsibleLintAnnotatorClipboardAction(private val ruleId: String) : AnsibleLintAnnotatorAction(message("action.copy-rule-id-to-clipboard")) {
     override fun invoke(project: Project, editor: Editor?, file: PsiFile?) {
         ClipboardSynchronizer
             .getInstance()
@@ -33,5 +17,9 @@ class AnsibleLintAnnotatorClipboardAction(private val ruleId: String) : Intentio
                 TextTransferable(ruleId, ruleId),
                 CopyPasteManagerEx.getInstanceEx()
             )
+    }
+
+    override fun getPriority(): Priority {
+        return Priority.LOW
     }
 }