+++ src/main/kotlin/de/achimonline/ansible_lint/settings/AnsibleLintSettings.kt
@@ -3,5 +3,6 @@ package de.achimonline.ansible_lint.settings
 data class AnsibleLintSettings(
     var executable: String = "ansible-lint",
     var offline: Boolean = false,
-    var onlyRunWhenConfigFilePresent: Boolean = true
+    var onlyRunWhenConfigFilePresent: Boolean = true,
+    var visualizeIgnoredRules: Boolean = true
 )